Modify zip and tar links to be less cryptic
The default github pages tar and zip links creates a download
name which includes the account owner and commit number, as in
nrc-cnrc-EGSnrc-2015-38-g4eb37cf for example. This cryptic
download name is bound to be confusing for users, so instead
we link to archive/master.zip and archive/master.tar.gz which
downloads to EGSnrc-master.zip (or .tar.gz).