Add an egs++ dynamic source class
Add an EGS_DynamicSource class to allow simulation of dynamic
(time-varying) motion of any source during a single calculation.
Continuous motion of the source is simulated between user-defined
control points, where each control point defines a set of incident polar
coordinates and an associated fractional monitor unit (mu) index.

If the source is an IAEA format phase space file or a BEAMnrc simulation
source which stores the value of mu for each particle, this mu value can
be used to set incident source coordinates, thus synchronizing the
motion of EGS_DynamicSource with jaw and MLC opening coordinates in the
upstream BEAMnrc simulation.

For phase space and BEAMnrc simulation sources, EGS_DynamicSource
essentially provides the same functionality as sources 20 and 21 in
DOSXYZnrc, as documented by Lobo & Popescu, Phys. Med. Biol.,
55:4431-4443 (2010).

Note that implementation of this source requires the implementation of a
new method, EGS_Float getMu(), in egs_base_source. As reimplemented in
iaea_phsp_source and egs_beam_source, getMu returns the value of mu
associated with a particle.

See the source documentation for more info and sample input.

Note: replaced to_string(), which is not a member of std on all g++
compilers, with another method of converting integers to strings during
the source input loop.