Reduce scope of radionuclide specific functions
Change the declaration of functions specific to radionuclide sources so
that they are not defined as virtual functions in the base source and
spectra classes. Instead, applications must use a static cast to a
radionuclide source in order to use those special functions.

Fix the implementation of radionuclide experiment time, so that it does
not crash when the time limit is hit. Now, source particles generated
after the time limit just return a zero energy particle to be ignored.

Add a radionuclide source function getEmissionType() that returns the
type of event that just occurred for the lastest source particle. This
can be used by applications to know what sort of particle has been
return or what event occurred (e.g. transition photon, Auger electon,
etc.)