Add app compile time and check for local differences
Add the compilation time of an application to the printout of git hash
and branch, and add these messages to the mortran codes.
Also add a check for applications in the makefile to compare with
the HEN_HOUSE versions and print a warning if there are differences.

With the git information now printed to machine.mortran, it is not
necessary to pass it as c++ compile flags. However, I will leave these
flags for future use.