Add a flag to specify if EM fields are on or off
Add the boolean variable emfield_on which is set to true if the
electromagnetic field (emf) is not null. This flag is used in the
emf_macros to skip all field related computations when there is no
field.

Ideally the file emf_macros.mortran would not be included at all in the
compilation when there is no field, but it proves impractical and error
prone to change the Makefile and recompile upon changing the input
parameters. The emfield_on flag allows the emf_macros to be integrated
by default, yet not waste time at each step when the field is 0.