Add target 'realclean' for all objects and libs
The 'clean' target only removes object files in the egs++ folder without
recursively going inside the folders for geometries, sources, etc. It is
meant to only rebuild the basic part of egs++.

Add a 'realclean' target which removes all objects files and libraries
from the dso/config folder that belong to the egs++ library including
sources and geometries. Useful when one wants to recompile the full
library.

Note that libiaea_phsp.so (iaea_phsp.lib on Windows) are preserved.