Fix #430: Fix blank inputs for DOSXYZnrc sources
Fixed a bug in inputs for sources 20, 21 in which the code
hung if the user indicated that a shared library geometry was not
to be included by inputting a blank (instead of a '0')
for the name of code defining the shared library geometry.

While fixing this, other input bugs related to incorrect
use of the Fortran len(char) were found and fixed.