Fix Qt GUI compile functions and tooltip
Fix a bug in the Qt GUI compilation functions. The make program from the
config may contain a space if it was provided arguments like -j12. Now
the Qt gui splits the string if there are spaces, so that it can
determine the make program correctly.

Also fix a bug on the odroid OS where the egs_inprz tooltip default
text color did not contrast with the tooltip background. This was fixed
by setting those colors explicitly.