Put distribution folder under git control
Tools for creating distribution files for the EGSnrc simulation system.

Currently only compressed and uncompressed tar-files are created for the
pre-compiled GUIs egs_gui, egs_inprz, and egs_view. The latter requires
the geometry, shape and track scoring modules to be included. GUIs are
linked statically with the Qt library for distribution. However it could
also be used with a dynamic Qt library as long as the targeted systems
use the same Qt and OS libraries.

The egs_configure GUI is also statically built and compressed *.egssself
archives appended to it for automatic deployment.

Build process on the Linux side is done directly, BUT one must have
first built the GUIs on a Windows machine. To do this, one must go to
$HEN_HOUSE/pieces/windows and run make (mingw32-make if using MinGW).

Once the Windows part is completed, one can from $HEN_HOUSE/distribution
finish the process by typing:

`make'

File egs_guis_list defines the name of the archive file and the list of
files to be included. It uses a special format which is described in the
header of egs_guis_list. Although currently only the GUIs and the egs++
geometry are included, one could create arbitrary archives, for
instance, with the documentation.