Fix #359: photon cs parameter not initialized
Fix the photon cross sections transport parameters initialization.
Although the correct default value was used when this parameter was not
specified, some corrupted text was output to the egslst file. This minor
bug did not affect results. Now, the variable is initialized in the same
way as the compton cross sections.